[Adapter] consolidate HTTP adapter
